Move module to a package fix for moreng ghc
