--- Playable.hs
@@ -1,12 +1,0 @@
-{-# LANGUAGE FlexibleInstances #-}
-module Playable(
-	  Playable(..)
-	) where 
-
-class Playable a where
-	sound :: a -> (Double,[Double])
-
-instance Playable (Double,[Double]) where 
-	sound = id 
-
-
