+++ stack.yaml
@@ -1,0 +1,10 @@
+
+resolver: lts-6.4
+
+packages:
+- '.'
+extra-deps: []
+
+flags: {}
+
+extra-package-dbs: []
